Fixed an issue with the symbols
The problem here was, that an external call of (day06:day06) did not produce the correct result.
While not strictly supported (the symbol 'day06 is not exported) this surprised me.
The issue was that for the second part of the solution, I try to lookup 2 symbols in the hash table. When run from another package the symbols withing the hash table would belong to the 'other' package while the code tries to lookup symbols from the 'day06'-package.
The fix is to make all symbols keywords, s.t. that they belong to the same package.